+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add support for global rules-exclude
+* Add `Halstead Volume` metric.
 
 ## 4.2.1